Pull helper code out from download.ts into separate files
