+++ lib/codeql.js
@@ -33,6 +33,7 @@ const exec_1 = require("@actions/exec");
 const yaml = __importStar(require("js-yaml"));
 const deserialize_1 = require("./deserialize");
 const download_1 = require("./download");
+const http_error_1 = require("./http-error");
 const json_validation_1 = require("./json-validation");
 const query_run_memory_1 = require("./query-run-memory");
 const query_run_metadata_1 = require("./query-run-metadata");
@@ -117,7 +118,7 @@ async function downloadDatabase(repoId, repoName, language, pat) {
     }
     catch (error) {
         console.log("Error while downloading database");
-        if (error instanceof download_1.HTTPError &&
+        if (error instanceof http_error_1.HTTPError &&
             error.httpStatusCode === 404 &&
             error.httpMessage.includes("No database available for")) {
             throw new Error(`Language mismatch: The query targets ${language}, but the repository "${repoName}" has no CodeQL database available for that language.`);
