+++ lib/download.js
@@ -25,30 +25,16 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.download = exports.HTTPError = exports.getApiClient = void 0;
+exports.download = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const stream = __importStar(require("stream"));
 const util = __importStar(require("util"));
 const core = __importStar(require("@actions/core"));
-const httpm = __importStar(require("@actions/http-client"));
 const io = __importStar(require("@actions/io"));
-const gh_api_client_1 = require("./gh-api-client");
-// TODO: replace this with octokit too
-function getApiClient() {
-    return new httpm.HttpClient(gh_api_client_1.userAgent, [], {
-        allowRetries: true,
-    });
-}
-exports.getApiClient = getApiClient;
-class HTTPError extends Error {
-    constructor(httpStatusCode, httpMessage) {
-        super(`Unexpected HTTP response: ${httpStatusCode}. ${httpMessage}`);
-        this.httpStatusCode = httpStatusCode;
-        this.httpMessage = httpMessage;
-    }
-}
-exports.HTTPError = HTTPError;
+const api_client_1 = require("./api-client");
+const http_error_1 = require("./http-error");
+const retry_helper_1 = require("./retry-helper");
 /**
  * Download a file from an url and stream it into a local file
  *
@@ -64,11 +50,11 @@ async function download(url, dest, auth, accept) {
     const maxAttempts = 3;
     const minSeconds = 10;
     const maxSeconds = 20;
-    const retryHelper = new RetryHelper(maxAttempts, minSeconds, maxSeconds);
+    const retryHelper = new retry_helper_1.RetryHelper(maxAttempts, minSeconds, maxSeconds);
     return await retryHelper.execute(async () => {
         return await downloadAttempt(url, dest, auth, accept);
     }, (err) => {
-        if (err instanceof HTTPError && err.httpStatusCode) {
+        if (err instanceof http_error_1.HTTPError && err.httpStatusCode) {
             // Don't retry anything less than 500, except 408 Request Timeout and 429 Too Many Requests
             if (err.httpStatusCode < 500 &&
                 err.httpStatusCode !== 408 &&
@@ -86,7 +72,7 @@ async function downloadAttempt(url, dest, auth, accept) {
         throw new Error(`Destination file path ${dest} already exists`);
     }
     // Get the response headers
-    const http = getApiClient();
+    const http = (0, api_client_1.getApiClient)();
     const headers = {};
     if (auth) {
         core.debug("set auth");
@@ -97,7 +83,7 @@ async function downloadAttempt(url, dest, auth, accept) {
     }
     const response = await http.get(url, headers);
     if (response.message.statusCode !== 200) {
-        const err = new HTTPError(response.message.statusCode, await response.readBody());
+        const err = new http_error_1.HTTPError(response.message.statusCode, await response.readBody());
         core.debug(`Failed to download from "${url}". Code(${err.httpStatusCode}) Message(${err.httpMessage})`);
         throw err;
     }
@@ -121,52 +107,6 @@ async function downloadAttempt(url, dest, auth, accept) {
                 core.debug(`Failed to delete '${dest}'. ${err instanceof Error ? err.message : err}`);
             }
         }
-    }
-}
-/**
- * Internal class for retries.
- * Borrowed from https://github.com/actions/toolkit/blob/main/packages/tool-cache/src/retry-helper.ts.
- */
-class RetryHelper {
-    constructor(maxAttempts, minSeconds, maxSeconds) {
-        if (maxAttempts < 1) {
-            throw new Error("max attempts should be greater than or equal to 1");
-        }
-        this.maxAttempts = maxAttempts;
-        this.minSeconds = Math.floor(minSeconds);
-        this.maxSeconds = Math.floor(maxSeconds);
-        if (this.minSeconds > this.maxSeconds) {
-            throw new Error("min seconds should be less than or equal to max seconds");
-        }
-    }
-    async execute(action, isRetryable) {
-        let attempt = 1;
-        while (attempt < this.maxAttempts) {
-            // Try
-            try {
-                return await action();
-            }
-            catch (err) {
-                if (!(err instanceof Error) || !isRetryable(err)) {
-                    throw err;
-                }
-                core.info(err.message);
-            }
-            // Sleep
-            const seconds = this.getSleepAmount();
-            core.info(`Waiting ${seconds} seconds before trying again`);
-            await this.sleep(seconds);
-            attempt++;
-        }
-        // Last attempt
-        return await action();
-    }
-    getSleepAmount() {
-        return (Math.floor(Math.random() * (this.maxSeconds - this.minSeconds + 1)) +
-            this.minSeconds);
-    }
-    async sleep(seconds) {
-        return new Promise((resolve) => setTimeout(resolve, seconds * 1000));
     }
 }
 //# sourceMappingURL=download.js.map
