+++ lib/query.js
@@ -40,7 +40,7 @@ async function run() {
     catch (error) {
         console.error(error);
         const errorMessage = error instanceof Error ? error.message : `${error}`;
-        if (error instanceof download_1.HTTPError && error.httpStatusCode === 403) {
+        if (error instanceof tool_cache_1.HTTPError && error.httpStatusCode === 403) {
             (0, core_1.setFailed)(`${errorMessage}. The query pack is only available for 24 hours. To retry, create a new variant analysis.`);
         }
         else {
@@ -68,7 +68,7 @@ async function run() {
         catch (error) {
             console.error(error);
             const errorMessage = error instanceof Error ? error.message : `${error}`;
-            if (error instanceof download_1.HTTPError && error.httpStatusCode === 403) {
+            if (error instanceof tool_cache_1.HTTPError && error.httpStatusCode === 403) {
                 (0, core_1.setFailed)(`${errorMessage}. Database downloads are only available for 24 hours. To retry, create a new variant analysis.`);
             }
             else {
