+++ src/api-client.ts
@@ -1,0 +1,9 @@
+import * as httpm from "@actions/http-client";
+
+import { userAgent } from "./gh-api-client";
+
+export function getApiClient(): httpm.HttpClient {
+  return new httpm.HttpClient(userAgent, [], {
+    allowRetries: true,
+  });
+}
