+++ src/http-error.ts
@@ -1,0 +1,9 @@
+export class HTTPError extends Error {
+  httpStatusCode: number | undefined;
+  httpMessage: string;
+  constructor(httpStatusCode: number | undefined, httpMessage: string) {
+    super(`Unexpected HTTP response: ${httpStatusCode}. ${httpMessage}`);
+    this.httpStatusCode = httpStatusCode;
+    this.httpMessage = httpMessage;
+  }
+}
